Create apiClients to isolate and formalize communication between backend and frontend [Re: nmkataoka/adv-life#69]
The api clients are modeled off of Socket.io in case we ever want to
add real sockets for server-client communication.

Character creation for player generation was moved into this new layer.

Hopefully this will work just as well if we decide to change the engine
to an immutable, flux-y data flow architecture.