Initialize inventory size [Re: nmkataoka/adv-life#69]
Players will have fixed inventory sizes, whereas non-player characters
will have unlimited inventories for now.

Removing items from inventories will leave an empty space in the
inventory both for performance reasons and so that player inventories
will 'remember' the empty spaces, which is the expected behavior in
inventory UI.

Optimization: the inventorycmpt could store the index of the next
empty slot and update when adding/deleting elements so that a full
search does not have to be run whenever items are added to the
inventory.