Add tests for DataDbCmpt and fix related bugs [Re: nmkataoka/adv-life#108]
Added tests for DataDbCmpt since it's a relatively complicated
container which will be reused a lot. As a consequence I found a bug
with the field initialization and constructor order. I misunderstood
the order of parent/child field initializers and child constructors.
Makes sense now why Loopback needs to have Models declare constructors
that seem to duplicate code.