Add ItemStackCmpt [Re: nmkataoka/adv-life#109]
Used ids instead of direct references for Material and ItemClass
because I'd like to standardize and use handles everywhere, instead of
pointer-like references which are more brittle and buggy.

Between ItemStackCmpt and the ItemClassDb, there won't actual be an
`item` class or `ItemCmpt`.

Still need to refactor all the code to use ItemStackCmpt.