Use a single generic param in component accessors [Re: nmkataoka/adv-life#115]
I created the generic signature for ComponentManager when I was still
pretty new to Typescript, and I couldn't figure out how to do it with
only 1 generic param (ended up using 2). The solution is pretty simple,
so I'm reducing it to 1 generic param and inferring the other one.
This greatly simplifies the generics around the component manager.