Update towns to use react-ecsal, use numbers for entities in View [Re: nmkataoka/adv-life#133]
Updated towns and town locations to use react-ecsal `useSelector`.
The game is currently bugged, likely because the update loop is no
longer triggering re-renders. Next commit will fix the Ecsal store
(GameManager) to trigger rerenders properly.

To get proper typing of Views, it seems like their types need to be
explicitly stated and can't be inferred from their constructors, even
with Typescript 4.0 variadic tuples. Unfortunately this means querying
using Views requires writing out the class lists for the View twice.

Updated View to use numbers for entities instead of strings. This has
an upfront cost of parsing the dictionary keys (strings) back into
numbers, but it seems like everyone expects numbers so it's much
more convenient for them to be numbers.

Removed TownScene test because we need to mock the Ecsal `Provider` for
tests that use Ecsal's `useSelector` and that's a low priority.