Update View to use ComponentDependencies to reduce boilerplate [Re: nmkataoka/adv-life#236]
Updated View and everything else in the engine to pass around
ComponentDependencies types instead of the individual component groups
ReadCmpts, WriteCmpts, WithoutCmpts.

This will be much more maintainable in the long run, and it reduces
code in the short run. All upsides, no downsides other than it's a
little more complex (uses a lot of type inference).