Add box blur and tweak noise parameters for fault perturbation [Re: nmkataoka/adv-life#305]
The noise parameters are not really set up properly to handle changing
the metersPerCoord property, so they needed to be tweaked since
the port involved changing meters per coord from 50k to 25k. When we
reduce to 4k, will need to do heavy tweaking on each terrain gen step
again.