Change metersPerCoord to 4092 [Re: nmkataoka/adv-life#302]
Changed default map metersPerCoord to 4092. I think I'll be sticking to
this for the foreseeable future. I did have to reduce the tec plates
to keep everything sane, but I think that's okay -- better to keep
measurements somewhat close to realistic (rather than divorce from
reality to make the smaller maps look like an Earth-sized planet) to
prevent difficulties in the future.