Noise t during fault propagation, tweak terrain gen defaults [Re: nmkataoka/adv-life#335]
Fixed some not great settings in shapeCoasts and lowFreqNoise. I like
what the defaults give you now except the mountains are a little short.
Not sure if this is the renderer or they're just short (I'd guess the
latter). We can continue tweaking the mountain ranges after we have
hydraulic erosion.